Baxter-King Filter

This function computes the cyclical component of the Baxter-King filter.

filter_bk(x, fill = NA, ...)

Arguments

x
[univariate vector]

Univariate vector, numeric or ts object with only one dimension.

fill
[numeric or function: NA]

Numeric value(s) or function used to fill observations.

...
Further arguments passed to bkfilter.

Examples

# \donttest{
x <- cumsum(rnorm(100))
filter_bk(x)

#>   [1]           NA           NA           NA -0.087049799 -0.542506723
#>   [6]  0.783385516  1.144147101 -0.030118577 -0.551260012 -0.925844429
#>  [11]  0.136403712  0.196890150  0.253939362 -0.006931137 -0.858135738
#>  [16]  0.731269448  1.151558040 -0.706579414  0.323706590  0.353398990
#>  [21] -0.153133581  0.257442353 -0.007734041  0.028921891  0.059354904
#>  [26] -1.101025941  0.067113015  0.314798192 -0.918856800  0.125584161
#>  [31]  0.243256082 -0.522652572 -0.214515464  0.017043219  0.160530666
#>  [36]  0.276833068  0.484624267  0.338767726  0.185374979  0.164701290
#>  [41] -0.149882802 -0.204384460 -1.570793772  0.355404241  1.339959609
#>  [46]  0.103551707 -0.263556166 -0.612060480  0.181130635 -0.001493795
#>  [51]  0.071566066 -0.194679580 -0.573962416  0.404451111 -0.163560957
#>  [56]  1.099019927 -0.589126413 -0.092035682 -0.050826974  0.099897019
#>  [61]  0.501721057  0.260271732  0.375080428 -0.146288628 -0.824847918
#>  [66] -0.364135264 -0.121212002 -0.593953640 -0.194805083  1.536114897
#>  [71]  0.981095777 -1.102894598  0.284882899 -0.115038650 -0.628734237
#>  [76]  0.560880286  0.491328090 -0.517811902 -0.138378733 -0.137245526
#>  [81] -0.127364867  0.177931840 -0.296775809  0.167801835 -0.338980401
#>  [86] -0.368642006  0.308673949  0.251195191 -0.646700841 -0.092535703
#>  [91]  0.356962823  0.410693015  0.265520519 -0.697073465  0.169294228
#>  [96] -1.130875484  0.307208118           NA           NA           NA

# }