Christiano-Fitzgerald Filter

This function computes the cyclical component of the Christiano-Fitzgerald filter.

filter_cf(x, ...)

Arguments

x
[univariate vector]

Univariate vector, numeric or ts object with only one dimension.

...
Further arguments passed to cffilter.

Examples

# \donttest{
x <- cumsum(rnorm(100))
filter_cf(x)

#>   [1] -0.512860344  0.034490666  0.187077096  0.312563016 -0.122423617
#>   [6]  0.378365350  0.164420007 -0.918873729 -0.736941189  0.658500043
#>  [11]  0.403101992  1.131096269 -0.550044906 -0.764895062 -0.365235014
#>  [16] -0.009847609  0.401866614  0.305192899  0.135769586 -0.219341824
#>  [21]  0.423844481 -0.207943365 -0.577283238 -0.834136918  0.966900591
#>  [26]  0.271439915  0.449366066  0.410542337 -0.757852378 -1.099011907
#>  [31]  0.103868621  0.462348311  0.643671040  0.583905025 -1.005948541
#>  [36]  0.499042483 -0.847487843 -0.307898126  1.197124664 -0.602836417
#>  [41]  0.086411360  0.364098536 -0.098253718 -0.135582210 -0.248976493
#>  [46]  0.341982027 -0.616426064 -0.077993831  1.404842267 -0.650861651
#>  [51] -0.470774633  0.038907337  0.466605315 -0.237256509 -0.050796030
#>  [56] -0.188826154  0.298687151 -0.307026828  0.399056462 -0.181172172
#>  [61]  0.758975412 -0.365916426 -1.764907474  1.194455768  0.344257397
#>  [66]  0.129032156  0.288112325 -0.537010171 -0.192433458  0.124460257
#>  [71] -0.123655416 -0.167507190 -0.424322962  1.410168555  0.403523917
#>  [76] -0.824555640 -0.731134059 -0.200610327  0.532610522  0.341443445
#>  [81] -0.555806644  0.766020468  0.418874760 -0.432230119 -0.607268965
#>  [86] -0.723525091  0.408678766  0.375404980  0.846232094 -0.031557056
#>  [91] -0.135855903 -0.540289397 -0.171587768 -0.464505953 -1.035213907
#>  [96]  1.528883339  2.194913856  0.283390803 -1.741395679 -4.902002766

# }