Hamilton Filter

This function computes the cyclical component of the Hamilton filter.

filter_hamilton(x, p = 4, horizon = 8, fill = NA)

Arguments

x
[univariate vector]

Univariate vector, numeric or ts object with only one dimension.

p
[integer(1): 4]

A value indicating the number of lags

horizon
[integer(1): 8]

A value indicationg the number of periods to look ahead.

fill
[numeric or function: NA]

Numeric value(s) or function used to fill observations.

Value

Returns a vector with the same class and attributes as the input vector.

Examples

x <- cumsum(rnorm(100))
filter_hamilton(x)

#>   [1]           NA           NA           NA           NA           NA
#>   [6]           NA           NA           NA           NA           NA
#>  [11]           NA  0.833688397 -0.779648562 -1.459882532 -1.443831632
#>  [16] -0.776045491 -1.035042398 -2.928987771 -1.627243316 -0.921572201
#>  [21] -0.243436433  1.215176166  0.107692815  0.527074821 -0.021293346
#>  [26]  0.877945238  0.546836900  0.986038600  2.070838692  1.890006541
#>  [31]  3.217944943  1.885393117  1.411402055  2.667421961  3.111985623
#>  [36]  0.853640182 -0.763529652 -1.038921441 -0.409391938 -0.185215440
#>  [41]  0.581527919  1.364977117  2.856628861  3.219569379  3.616328893
#>  [46]  2.070985526  1.901403401  1.234857342  0.073090437 -0.262920068
#>  [51]  0.326287038 -1.725893277 -2.151504260 -1.533339910 -2.390760187
#>  [56] -1.857078444 -1.353539774 -1.285155020 -3.961407586 -1.052583524
#>  [61] -1.169440636 -0.361474180 -0.025776695  0.915343065  1.685426603
#>  [66]  1.388811568  2.305856299  0.908317267  1.849601092  0.949159579
#>  [71]  3.488507482  1.562173050  0.927345287  1.746240736  2.218072872
#>  [76]  1.892473830  0.704434454  0.996556718  0.362842617 -1.055200937
#>  [81] -1.021357927 -0.738542946 -0.798375442 -1.814291493 -1.068742867
#>  [86]  0.373155384  0.831754092  0.002475584 -0.377541440  0.010154154
#>  [91] -0.814592165 -2.755845434 -1.938129580 -3.016219015 -3.828946024
#>  [96] -2.122862311 -2.689429961 -1.818273648 -3.012827189 -2.931349604