Hodrick-Prescot Filter

This function computes the cyclical component of the Hodrick-Prescot filter.

filter_hp(x, ...)

Arguments

x
[univariate vector]

Univariate vector, numeric or ts object with only one dimension.

...
Further arguments passed to hpfilter.

See also

Examples

# \donttest{
x <- cumsum(rnorm(100))
filter_hp(x)

#>   [1]  1.67956966  0.60742292  0.06812960  0.13226389  0.89198466 -0.67504276
#>   [7] -0.94889565 -0.12589585 -0.60851676 -0.33646019  0.18847132  0.47651969
#>  [13]  1.13695551  1.00754630  0.57173649 -2.11116229 -2.26401080 -1.91501289
#>  [19] -1.48208676 -2.15963351 -0.51987812 -0.38701403 -0.42408006  0.67921640
#>  [25] -0.20931038 -0.82672615  1.41083711  1.27398617  2.77532300  1.22337647
#>  [31]  0.93952010  1.24524872  1.49319309  0.45524225  0.46021182 -0.61551294
#>  [37]  0.11254500  1.22439123 -0.96382136  0.31557342 -0.87701896 -0.37141940
#>  [43]  0.33986721  0.19032269 -0.40728941 -0.19923368  0.27588339  1.18750361
#>  [49] -0.84621402 -2.47613124 -1.05335454 -0.02882183  0.37017606  1.03516203
#>  [55]  1.88830535 -0.85107299  0.16455966 -0.43249432 -0.33270172 -0.77850717
#>  [61] -0.07787094 -0.61872425 -0.31397591 -0.93972657 -2.28862919 -1.35493596
#>  [67] -0.90827686  0.50694593  0.25195464  1.04908230  2.69211486  2.07629760
#>  [73]  0.41794290 -0.15829819 -0.71139359 -0.90151164  0.46611523 -0.20715801
#>  [79] -0.18359697 -0.76277248 -1.09547982 -1.13538890 -0.16040702 -0.38933531
#>  [85] -0.02140103  0.42048540  1.00882871  0.12089421  1.44884136  0.78519676
#>  [91]  0.41370572  1.56927412  2.63208666  1.32703857  0.25248614 -1.29696935
#>  [97] -1.30002116 -1.31641887 -1.13188677 -0.75883738

select_lambda("monthly")

#> [1] 14400

filter_hp(x, freq = 100)

#>   [1]  0.872884034 -0.026817392 -0.401344498 -0.187154400  0.704936492
#>   [6] -0.748427534 -0.934836704 -0.051063411 -0.491276332 -0.186440552
#>  [11]  0.374910493  0.717942003  1.463216495  1.452911842  1.160961245
#>  [16] -1.376734288 -1.417698401 -1.012349988 -0.588060312 -1.339756899
#>  [21]  0.164763116  0.117781963 -0.129236080  0.747085000 -0.371329827
#>  [26] -1.214481923  0.812143558  0.501428379  1.881009788  0.270233926
#>  [31] -0.015783958  0.336255002  0.671530582 -0.248103371 -0.109648533
#>  [36] -1.049771347 -0.195662402  1.031129307 -1.052770928  0.311229386
#>  [41] -0.815612609 -0.265181224  0.475563234  0.348066835 -0.231485815
#>  [46] -0.009300005  0.478137425  1.402356377 -0.621004987 -2.258610625
#>  [51] -0.871688070  0.099742253  0.447365498  1.080665160  1.935528771
#>  [56] -0.765164063  0.311502809 -0.209665016 -0.029643263 -0.399559186
#>  [61]  0.364032839 -0.131883516  0.191008869 -0.451232631 -1.861209251
#>  [66] -1.039196470 -0.743665839  0.501684001  0.085647805  0.753431142
#>  [71]  2.320993478  1.698874997  0.097155596 -0.381437207 -0.818505211
#>  [76] -0.893135207  0.578262378 -0.005767605  0.084207836 -0.459755321
#>  [81] -0.797782412 -0.888752533 -0.008489058 -0.365531268 -0.148859937
#>  [86]  0.132272977  0.565505879 -0.457822147  0.763497572  0.035694625
#>  [91] -0.350512131  0.846890554  2.018958391  0.893969563  0.055118739
#>  [96] -1.226264689 -0.952539277 -0.695667818 -0.244874996  0.390060252

# }