Rolling operations

Apply rolling operations over a moving window for size n and increment step.

roll_idx(
  x,
  n = 1,
  step = 1,
  align = c("left", "center", "right"),
  complete = TRUE
)

roll(
  x,
  fn,
  n = 1L,
  step = 1L,
  fill = NA,
  align = c("left", "center", "right"),
  ...
)

roll_data(x, n = 1L, step = 1L, align = c("left", "center", "right"))

Arguments

x
[univariate vector]

Univariate vector, numeric or ts object with only one dimension.

n
[positive integer(1):1L]

Window size.

step
[positive integer(1):1L]

Rolling window step.

align
[character(1): "left"]

Specifying whether the index of the result should be left- or right-aligned or centered (default) compared to the rolling window of observations.

fn
[function]

fill
[numeric or function: NA]

Numeric value(s) or function used to fill observations.

...
Further arguments passed to the function fn.

Value

roll(): returns a vector with the same class and attributes as the input vector.

roll_idx(): Returns the index that is used to calculate the subsets.

roll_data(): Returns a list subsets of length length(x)/step.

Examples

# x is odd

roll_idx(1:9, 2, 1, align = "left")

#> Error in roll_idx(1:9, 2, 1, align = "left"): could not find function "roll_idx"

roll_idx(1:9, 2, 1, align = "center")

#> Error in roll_idx(1:9, 2, 1, align = "center"): could not find function "roll_idx"

roll_idx(1:9, 2, 1, align = "right") # reduces

#> Error in roll_idx(1:9, 2, 1, align = "right"): could not find function "roll_idx"

# This works
roll_idx(1:9, 3, 1, align = "left")

#> Error in roll_idx(1:9, 3, 1, align = "left"): could not find function "roll_idx"

roll_idx(1:9, 3, 1, align = "center")

#> Error in roll_idx(1:9, 3, 1, align = "center"): could not find function "roll_idx"

roll_idx(1:9, 3, 1, align = "right")

#> Error in roll_idx(1:9, 3, 1, align = "right"): could not find function "roll_idx"

roll_idx(1:9, 2, 2, align = "left")

#> Error in roll_idx(1:9, 2, 2, align = "left"): could not find function "roll_idx"

roll_idx(1:9, 2, 2, align = "center")

#> Error in roll_idx(1:9, 2, 2, align = "center"): could not find function "roll_idx"

roll_idx(1:9, 2, 2, align = "right")

#> Error in roll_idx(1:9, 2, 2, align = "right"): could not find function "roll_idx"

roll_idx(1:9, 3, 2, align = "left")

#> Error in roll_idx(1:9, 3, 2, align = "left"): could not find function "roll_idx"

roll_idx(1:9, 3, 2, align = "center")

#> Error in roll_idx(1:9, 3, 2, align = "center"): could not find function "roll_idx"

roll_idx(1:9, 3, 2, align = "right") # reduces

#> Error in roll_idx(1:9, 3, 2, align = "right"): could not find function "roll_idx"

# x is even

roll_idx(1:8, 2, 2, align = "left")

#> Error in roll_idx(1:8, 2, 2, align = "left"): could not find function "roll_idx"

roll_idx(1:8, 2, 2, align = "center")

#> Error in roll_idx(1:8, 2, 2, align = "center"): could not find function "roll_idx"

roll_idx(1:8, 2, 2, align = "right") # reduces

#> Error in roll_idx(1:8, 2, 2, align = "right"): could not find function "roll_idx"

roll_idx(1:8, 2, 1, align = "left")

#> Error in roll_idx(1:8, 2, 1, align = "left"): could not find function "roll_idx"

roll_idx(1:8, 2, 1, align = "center")

#> Error in roll_idx(1:8, 2, 1, align = "center"): could not find function "roll_idx"

roll_idx(1:8, 2, 1, align = "right")

#> Error in roll_idx(1:8, 2, 1, align = "right"): could not find function "roll_idx"

roll_idx(1:8, 3, 1, align = "left")

#> Error in roll_idx(1:8, 3, 1, align = "left"): could not find function "roll_idx"

roll_idx(1:8, 3, 1, align = "center")

#> Error in roll_idx(1:8, 3, 1, align = "center"): could not find function "roll_idx"

roll_idx(1:8, 3, 1, align = "right") # reduces

#> Error in roll_idx(1:8, 3, 1, align = "right"): could not find function "roll_idx"

roll_idx(1:8, 3, 1, align = "left")

#> Error in roll_idx(1:8, 3, 1, align = "left"): could not find function "roll_idx"

roll_idx(1:8, 3, 1, align = "center")

#> Error in roll_idx(1:8, 3, 1, align = "center"): could not find function "roll_idx"

roll_idx(1:8, 3, 1, align = "right") # reduces

#> Error in roll_idx(1:8, 3, 1, align = "right"): could not find function "roll_idx"

x <- seq(10, 1, -1)

roll_data(x, 2, align = "left")

#> [[1]]
#> [1] 10  9
#> 
#> [[2]]
#> [1] 9 8
#> 
#> [[3]]
#> [1] 8 7
#> 
#> [[4]]
#> [1] 7 6
#> 
#> [[5]]
#> [1] 6 5
#> 
#> [[6]]
#> [1] 5 4
#> 
#> [[7]]
#> [1] 4 3
#> 
#> [[8]]
#> [1] 3 2
#> 
#> [[9]]
#> [1] 2 1
#> 
#> [[10]]
#> [1]  1 NA
#> 

roll_data(x, 2, align = "right")

#> [[1]]
#> [1] 9 8
#> 
#> [[2]]
#> [1] 8 7
#> 
#> [[3]]
#> [1] 7 6
#> 
#> [[4]]
#> [1] 6 5
#> 
#> [[5]]
#> [1] 5 4
#> 
#> [[6]]
#> [1] 4 3
#> 
#> [[7]]
#> [1] 3 2
#> 
#> [[8]]
#> [1] 2 1
#> 
#> [[9]]
#> [1]  1 NA
#> 

roll(x, max, 3)

#>  [1] 10  9  8  7  6  5  4  3 NA NA

roll(x, max, 3, align = "right")

#>  [1] NA NA 10  9  8  7  6  5  4  3 NA NA

x <- 1:6
roll_data(x, 2)

#> [[1]]
#> [1] 1 2
#> 
#> [[2]]
#> [1] 2 3
#> 
#> [[3]]
#> [1] 3 4
#> 
#> [[4]]
#> [1] 4 5
#> 
#> [[5]]
#> [1] 5 6
#> 
#> [[6]]
#> [1]  6 NA
#> 

roll(x, mean, 2)

#> [1] 1.5 2.5 3.5 4.5 5.5  NA

roll_data(x, 2, 2)

#> [[1]]
#> [1] 1 2
#> 
#> [[2]]
#> [1] 3 4
#> 
#> [[3]]
#> [1] 5 6
#> 

roll(x, mean, 2, 2)

#> [1] 1.5  NA 3.5  NA 5.5  NA