Christiano-Fitzgerald Filter

This function computes the cyclical component of the Christiano-Fitzgerald filter.

filter_cf(x, ...)

Arguments

x
[univariate vector]

Univariate vector, numeric or ts object with only one dimension.

...
Further arguments passed to cffilter.

Examples

# \donttest{
x <- cumsum(rnorm(100))
filter_cf(x)

#>   [1]  0.200411036  0.830342788  0.092543419  0.423974328  0.050369682
#>   [6] -0.193616049 -0.660532763 -0.971389949  0.797128840  0.600256685
#>  [11]  0.294329291 -0.092832646  0.473287606 -0.526466712 -1.630455715
#>  [16]  0.387715721  0.671802749  0.847911925  0.414222801 -0.389474533
#>  [21] -0.892706655 -0.535578528  0.183222925  0.593842270  0.167788937
#>  [26]  0.395527160 -0.045185328 -0.586661788  0.256033877 -1.021706781
#>  [31]  0.579411259  0.235861849  0.457009425  0.031078490 -0.008749860
#>  [36] -0.733039675 -0.421121666  0.321066259  0.826237736  0.174229300
#>  [41] -0.458409538 -0.773643201  0.865680567 -0.271764900 -0.637008509
#>  [46]  0.964748831  0.624894560 -0.713114908 -0.988171337  0.186276311
#>  [51]  0.554984716  0.470081119  0.086555873 -0.422525014  0.220277059
#>  [56] -0.807570947 -0.275106190  0.398551437  0.846431805  0.077706142
#>  [61]  0.109598007 -0.358814651 -0.904293565 -0.626605113  0.568139250
#>  [66]  1.294585294  0.810857882 -0.862543102 -0.956271695 -0.637499984
#>  [71]  0.180182336  0.835017734  1.062869872 -0.741306517 -0.103358279
#>  [76] -0.531920966 -0.190518034  0.070766504  0.612554731  0.667298528
#>  [81] -0.994873239 -0.211003524  0.273809544  0.085488208  0.139051100
#>  [86] -0.027803267 -0.346988981  0.614607165 -0.274655039 -0.665773394
#>  [91]  0.035286504  0.570282774  0.959754376 -0.701479314  0.004371081
#>  [96] -1.095900683  0.203069512 -0.709376939  0.907301249  4.700130025

# }