Hamilton Filter

This function computes the cyclical component of the Hamilton filter.

filter_hamilton(x, p = 4, horizon = 8, fill = NA)

Arguments

x
[univariate vector]

Univariate vector, numeric or ts object with only one dimension.

p
[integer(1): 4]

A value indicating the number of lags

horizon
[integer(1): 8]

A value indicationg the number of periods to look ahead.

fill
[numeric or function: NA]

Value(s) of function used to fill observations.

Value

Returns a vector with the same class and attributes as the input vector.

Examples

x <- cumsum(rnorm(100))
filter_hamilton(x)

#>   [1]          NA          NA          NA          NA          NA          NA
#>   [7]          NA          NA          NA          NA          NA -0.18272501
#>  [13]  0.09368500 -0.65718549 -1.27904790  1.53528840  2.74737402  1.10013764
#>  [19]  0.55074780 -0.16675319 -1.56599372 -1.51886527 -2.03938539 -4.05018568
#>  [25] -4.75363607 -5.05758723 -4.59678324 -3.89861307 -4.78274011 -3.36808293
#>  [31] -2.34829467 -2.51456178 -1.33179342  0.43323952  0.22749633  0.76903407
#>  [37]  1.78536680  0.84439209  0.43914038  0.08073488 -1.01137990 -1.62815995
#>  [43] -3.42181229 -1.41716627 -0.29972163 -1.10105485 -1.12721385 -1.23495147
#>  [49] -0.02406352 -0.13226106  0.91365625 -1.03381231 -1.69326541  0.38919050
#>  [55]  0.80386903  2.87079091  0.42514869  1.08694136  0.91307016  1.36692164
#>  [61]  1.75612727  0.31456816  0.33786875 -1.81137248 -1.22403630 -1.61056823
#>  [67] -0.82729945 -1.26718724 -0.45184303  2.04804761  1.77813696  0.24226850
#>  [73]  1.75654911  0.58552934 -0.62759920  0.17888280 -0.69233705 -3.15917474
#>  [79] -2.30306833 -0.61060033 -1.19919067 -0.28544821 -0.78881653 -0.76030403
#>  [85] -0.79462777  0.22850490  1.25127884  1.73781874  1.10575491  2.01722883
#>  [91]  3.30337150  3.33520121  3.91419753  2.93287997  3.69507165  2.89125365
#>  [97]  5.46393947  6.37451572  5.68520245  4.34014690